[spaces] moved ContinuousLagrange -> CG, refs #14
